code/fastapi-admin/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
psutil>=5.9.0
PyJWT>=2.8.0
